fix: 스켈레톤 UI 렌더링 버그 수정
