+++ app/detail/[productId]/page.tsx
@@ -12,11 +12,7 @@ export default async function Page({
   params: Promise<{ productId: string }>;
 }) {
   const resolvedParams = await params;
-  console.log("전체 params 값:", resolvedParams);
-  // export default async function Page({ params }: PageProps) {
-  // console.log("params.productId 값", params.productId); //undefined
   const productId = Number(resolvedParams.productId); // URL에서 productId 가져오기
-  console.log("변환된 productId 값", productId); //NaN
 
   // 서버에서 리뷰 데이터를 가져옴
   const reviewData = await fetchReviews(productId);
@@ -25,8 +21,6 @@ export default async function Page({
     console.log("reviewData:", reviewData);
     return <p>리뷰 데이터를 불러오는 중 오류가 발생했습니다.</p>;
   }
-
-  console.log("reviewData.reviews:", reviewData.reviews);
 
   return (
     <ProductDetail productId={productId} initialReviews={reviewData.reviews} />
